* commit 'cea2f14661701feabf175ac9c172362962cf5a3e':
  [refactor] Oops, replace patched extern files
  [refactor] Add extraData field at then end of the message struct so we do not anymore "msg+1" playing with pointer to place data next to the struct
  [refactor] Cleaning some -Wall warnings
  [refactor] renambe clientId to lfClientId (compared to tcpClientId) in message header. And change type to fixed size (uint64_t)
  [refactor] Change msgType to uint64_t to get fixed size, also rename it from 'type'. Plus minor warning fixes

+++ src/base/network/HookLambdaFunction.cpp
@@ -28,7 +28,7 @@ HookLambdaFunction::~HookLambdaFunction(void)
 
 /****************************************************/
 //inherit
-LibfabricActionResult HookLambdaFunction::onMessage(LibfabricConnection * connection, int lfClientId, size_t msgBufferId, LibfabricMessage * clientMessage)
+LibfabricActionResult HookLambdaFunction::onMessage(LibfabricConnection * connection, uint64_t lfClientId, size_t msgBufferId, LibfabricMessage * clientMessage)
 {
 	return this->function(connection, lfClientId, msgBufferId, clientMessage);
 }
